Add lock with precedence to the MemState.
And use this instead of the global variable MVar.

We use this lock with low precedence for transactions, and with high
precedence for blocks.  In addition, when adding a block, we take the
lock before modifying the database, and release it only after both the
database and the mempool have been updated.  This way, we prevent
transations from being rejected due to a mismatch (CSL-1533).